Reworked the build script a bit
Renamed it from external/build_dependencies.sh to build_all.sh as
it makes little sense to keep it inside the external subdirectory.

Also, modified the build script so that it can be called from
anywhere, and that it does not move around with pushd/popd anymore,
but rather gives proper full paths to cmake and make commnads.